webviewtracer-crawler/scripts/requirements.txt
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.5
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymongo==4.3.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
requests==2.28.2
requests-file==2.1.0
rich==13.3.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tabulate==0.9.0
tldextract==5.3.0
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
